% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.gene.type.R
\name{classify.gene.type}
\alias{classify.gene.type}
\title{Gene type classification}
\usage{
classify.gene.type(smp.eff, ary.eff, smp.eff.train.ind, ary.eff.train.ind,
  group.id, ary.to.smp.assign)
}
\arguments{
\item{smp.eff}{estimated sample effect data, rows as probes, columns as samples; can only take in either probe-level data with 10 probe per unique probe or probe-set-level data.}

\item{ary.eff}{estimated array effect data, rows as probes, columns as samples; must have same dimensions and same probe name as sample effect data; can only take in either probe-level data with 10 probe per unique probe or probe-set-level data; must be the same dimensions as the estimated sample effect data.}

\item{smp.eff.train.ind}{training set index for samples from the estimated sample effect data.}

\item{ary.eff.train.ind}{training set index for arrays from the estimaed array effect data.}

\item{group.id}{sample group ID for the estimated sample effect data.}

\item{ary.to.smp.assign}{array-to-sample assignment for the arrays of the estimated array effect data.}
}
\value{
gene category vector, -1 for technical, 0 for other, 1 for biological genes
}
\description{
Classifies genes into technical, biological or other based on the differential expression analysis results of the estimated sample and array effect data.
}
\examples{
\dontrun{
smp.eff <- estimate.smp.eff(r.data = r.data.pl)
ary.eff <- estimate.ary.eff(r.data = r.data.pl,
                            non.r.data = non.r.data.pl)
ctrl.genes <- unique(rownames(r.data.pl))[grep("NC", unique(rownames(r.data.pl)))]
smp.eff.nc <- smp.eff[!rownames(smp.eff) \%in\% ctrl.genes, ]
ary.eff.nc <- ary.eff[!rownames(ary.eff) \%in\% ctrl.genes, ]
group.id <- substr(colnames(smp.eff.nc), 7, 7)
smp.eff.train.ind <- colnames(smp.eff.nc)[c(sample(which(group.id == "E"), size = 64),
sample(which(group.id == "V"), size = 64))]
smp.eff.test.ind <- colnames(smp.eff.nc)[!colnames(smp.eff.nc) \%in\% smp.eff.train.ind]
ary.eff.train.ind <- colnames(ary.eff.nc)[c(1:64, 129:192)]
group.id.list <- list("all" = group.id,
                      "tr" = substr(smp.eff.train.ind, 7, 7),
                      "te" = substr(smp.eff.test.ind, 7, 7))
ary.to.smp.assign <- list("all" = c(rep(c("E", "V"), each = 64),
                          rep(c("V", "E"), each = 32)),
                          "tr" = rep(c("E", "V"), each = 64),
                          "te" = rep(c("V", "E"), each = 32))
gene.cat <- classify.gene.type(smp.eff = smp.eff.nc,
                               ary.eff = ary.eff.nc,
                               smp.eff.train.ind = smp.eff.train.ind,
                               ary.eff.train.ind = ary.eff.train.ind,
                               group.id = group.id.list,
                               ary.to.smp.assign = ary.to.smp.assign)
}
}
\keyword{classification}

