% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounding.design.R
\name{confounding.design}
\alias{confounding.design}
\title{Confounding Design}
\usage{
confounding.design(seed, num.smp, degree = "complete", rev.order = FALSE)
}
\arguments{
\item{seed}{specifies seed for random assignment using set.seed().}

\item{num.smp}{number of samples.}

\item{degree}{level of confounding; has to be either "complete" or "partial" for either "complete confounding" or "partial confounding" design; default is "complete".}

\item{rev.order}{FALSE indicating no reverse order, first half arrays to group 1 (endometrial), second half arrays to group 2 (ovarian); default is FALSE.}
}
\value{
array-to-sample assignment, first half for group 1 (endometrial), second half for group 2 (ovarian)
}
\description{
Assigns arrays to samples with confounding design.
}
\examples{
cc.ind <- confounding.design(seed = 1, num.smp = 128,
                             degree = "complete", rev.order = FALSE)
cc.ind <- confounding.design(seed = 1, num.smp = 128,
                             degree = "complete", rev.order = FALSE)
}
\keyword{study.design}

