% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.smp.eff.R
\name{estimate.smp.eff}
\alias{estimate.smp.eff}
\title{Estimated Sample Effects}
\usage{
estimate.smp.eff(uhdata)
}
\arguments{
\item{uhdata}{the uniformly-handled expression dataset.
The dataset must have rows as probes and columns as samples.}
}
\value{
an estimation of the sample effects
}
\description{
Estimate sample effects from the expressions of the uniformly-handled data.
}
\examples{
smp.eff <- estimate.smp.eff(uhdata = uhdata.pl)

}
\keyword{data.setup}

