% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.pbset.R
\name{limma.pbset}
\alias{limma.pbset}
\title{Differential expression analysis of probe-set data}
\usage{
limma.pbset(data, group.id, group.id.level = c("E", "V"), pbset.id = NULL)
}
\arguments{
\item{data}{expression dataset to be differentially expression analyzed, rows as unique probe-sets, columns as samples.}

\item{group.id}{sample group label; must be a 2-level non-numeric factor vector.}

\item{group.id.level}{sample group label level, the first one being the reference level; default = c("E", "V") in our studies when comparing endometrial to ovarian samples.}

\item{pbset.id}{unique probe-set name; default is NULL, the rownames of the dataset.}
}
\value{
differential expression anlysis results, group means, group standard deviations
}
\description{
Performs two-group differential expression analysis using "limma".
}
\examples{
r.data.psl <- med.sum.pbset(data = r.data.pl,
                            num.per.unipbset = 10)
ctrl.genes <- unique(rownames(r.data.pl))[grep("NC", unique(rownames(r.data.pl)))]
r.data.psl.nc <- r.data.psl[!rownames(r.data.psl) \%in\% ctrl.genes, ]
group.id <- substr(colnames(r.data.psl.nc), 7, 7)
group.id.level <- levels(as.factor(group.id))
limma.fit.r.data<- limma.pbset(data = r.data.psl.nc,
                               group.id = group.id,
                               group.id.level = group.id.level)
                               table(limma.fit.r.data$P.Value < 0.01, dnn = "DE genes")
}
\keyword{DEA}

