% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.norm.R
\name{med.norm}
\alias{med.norm}
\title{Median nomalization}
\usage{
med.norm(train, test = NULL)
}
\arguments{
\item{train}{training data to be median normalized, rows as probes, columns as samples.}

\item{test}{test data to be frozen median normalized, rows as probes with equal number of rows as the training set, columns as samples.}
}
\value{
a list of two datasets, the normalized training set and the frozen normalied test set
}
\description{
Normalizes training dataset so that each array shares a same median, stores the median from the training dataset as the reference to frozen median normalize test dataset.
}
\examples{
set.seed(101)
group.id <- substr(colnames(non.r.data.pl), 7, 7)
train.ind <- colnames(non.r.data.pl)[c(sample(which(group.id == "E"), size = 64),
                               sample(which(group.id == "V"), size = 64))]
train.dat <- non.r.data.pl[, train.ind]
test.dat <- non.r.data.pl[, !colnames(non.r.data.pl) \%in\%train.ind]
data.mn <- med.norm(train = train.dat)
str(data.mn)
data.mn <- med.norm(train = train.dat, test = test.dat)
str(data.mn)
}
\keyword{preprocess}

