% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.sum.pbset.R
\name{med.sum.pbset}
\alias{med.sum.pbset}
\title{Probe-set median summarization}
\usage{
med.sum.pbset(data, pbset.id = NULL, num.per.unipbset = 10)
}
\arguments{
\item{data}{expression data to be summarized, rows as probes, columns as samples; rownames as probe names;
only accept data matrix with a fixed number of probes per unique probe-set. If the data is already on the probe-set level, no manipulation will be done.}

\item{pbset.id}{unique probe-set name, if not specified then use the unique probe name of the data.}

\item{num.per.unipbset}{number of probes for each unique probe-set; default is 10.}
}
\value{
probe-set median summarized data
}
\description{
Summarizes probe-set using median of each unique probe, only taking in data matrix with a fixed number of probes per unique probe-set.
}
\examples{
r.data.psl <- med.sum.pbset(data = r.data.pl,
                            num.per.unipbset = 10)
}
\keyword{preprocess}

