% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{non.r.data.pl}
\alias{non.r.data.pl}
\title{The non-randomized (test) probe-level dataset, 10 probes for each unique probe}
\format{A data matrix with 1810 rows (probes) and 192 columns (samples), column names ending with E/V are endometrial/ovarian samples.}
\usage{
non.r.data.pl
}
\description{
The non-randomized probe-level dataset, non-control-probe-removed, 10 probes for each unique probe, no background adjusted and after logged 2.
}
\keyword{example.data}

