% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.norm.R
\name{quant.norm}
\alias{quant.norm}
\title{Quantile normalization}
\usage{
quant.norm(train, test = NULL)
}
\arguments{
\item{train}{training dataset to be quantile normalized.
The dataset must have rows as probes and columns as samples.}

\item{test}{test dataset to be frozen quantile normalized.
The dataset must have rows as probes and columns as samples.
The number of rows must equal to the number of rows in the training set.
By default, the test set is not specified (\code{test = NULL}) and no frozen normalization will be performed.}
}
\value{
a list of two datasets:
\item{train.mn}{the normalized training set}
\item{test.fmn}{the frozen normalized test set, if test set is specified}
}
\description{
Normalize training dataset with quantile normalization and
store the quantiles from the training dataset as the references to frozen quantile normalize test dataset.
}
\examples{
set.seed(101)
group.id <- substr(colnames(non.r.data.pl), 7, 7)
train.ind <- colnames(non.r.data.pl)[c(sample(which(group.id == "E"), size = 64),
                               sample(which(group.id == "V"), size = 64))]
train.dat <- non.r.data.pl[, train.ind]
test.dat <- non.r.data.pl[, !colnames(non.r.data.pl) \%in\%train.ind]
data.qn <- quant.norm(train = train.dat)
str(data.qn)
data.qn <- quant.norm(train = train.dat, test = test.dat)
str(data.qn)
}
\references{
Bolstad, B. M., Irizarry R. A., Astrand, M, and Speed, T. P. (2003)
A Comparison of Normalization Methods for High Density Oligonucleotide Array Data Based on Bias and Variance.
Bioinformatics 19(2) , pp 185-193.
http://bmbolstad.com/misc/normalize/normalize.html
}
\keyword{preprocess}

