% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.description.R
\docType{data}
\name{r.data.pl}
\alias{r.data.pl}
\title{The uniformly-handled (benchmark) probe-level dataset, 10 probes for each unique probe}
\format{A data matrix with 1810 rows (probes) and 192 columns (samples).
Its column names ending with "E" or "V" indicate whether a sample belongs to endometrial or ovarian tumor sample.}
\usage{
r.data.pl
}
\description{
The uniforly-handled probe-level dataset, with non-control-probe removed,
10 probes per each unique probe, no background adjustment.
The expressions are on a log-2 scale.
}
\keyword{example.data}

