% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.signal.R
\name{reduce.signal}
\alias{reduce.signal}
\title{Biological signal reduction}
\usage{
reduce.signal(smp.eff, group.id, group.id.level = c("E", "V"),
  reduce.multiplier = 1/2, pbset.id = NULL)
}
\arguments{
\item{smp.eff}{estimated sample effect data, rows as probes, columns as samples; can only take in probe-level data with 10 probe per unique probe.}

\item{group.id}{sample group ID for the estimated sample effect data.}

\item{group.id.level}{sample group label level; default = c("E", "V") in our studies when comparing endometrial to ovarian samples.}

\item{reduce.multiplier}{a multiplier specified to reduce between-sample-group signal by; default is 1/2.}

\item{pbset.id}{unique probe-set name, if not specified then use the unique probe name of the data.}
}
\value{
estimated sample effect data with reduced biological signal
}
\description{
Reduces biological effect between sample group by a multiplier.
}
\examples{
smp.eff <- estimate.smp.eff(r.data = r.data.pl)
ary.eff <- estimate.ary.eff(r.data = r.data.pl,
                             non.r.data = non.r.data.pl)
ctrl.genes <- unique(rownames(r.data.pl))[grep("NC", unique(rownames(r.data.pl)))]
smp.eff.nc <- smp.eff[!rownames(smp.eff) \%in\% ctrl.genes, ]
ary.eff.nc <- ary.eff[!rownames(ary.eff) \%in\% ctrl.genes, ]
group.id <- substr(colnames(smp.eff.nc), 7, 7)
redhalf.smp.eff.nc <- reduce.signal(smp.eff = smp.eff.nc,
                                    group.id = group.id,
                                    group.id.level = c("E", "V"),
                                    reduce.multiplier = 1/2)
}
\keyword{data.setup}

