% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.design.R
\name{stratification.design}
\alias{stratification.design}
\title{Stratification Design}
\usage{
stratification.design(seed, num.smp, batch.id)
}
\arguments{
\item{seed}{specifies seed for random assignment using set.seed().}

\item{num.smp}{number of samples.}

\item{batch.id}{sample group ID for the estimated sample effect data.}
}
\value{
array-to-sample assignment, first half for group 1 (endometrial), second half for group 2 (ovarian)
}
\description{
Assigns arrays to samples with stratification design.
}
\examples{
batch.id <- list(1:40, 41:64, (129:160) - 64, (161:192) - 64)
str.ind <- stratification.design(seed = 1, num.smp = 128,
                                 batch.id = batch.id)
}
\keyword{study.design}

