% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni.handled.simulate.R
\name{uni.handled.simulate}
\alias{uni.handled.simulate}
\title{Classification analysis of uniformly-handled data}
\usage{
uni.handled.simulate(myseed, N, smp.eff, norm.list = c("NN", "QN"),
  class.list = c("PAM", "LASSO"), norm.funcs = NULL, class.funcs = NULL,
  pred.funcs = NULL)
}
\arguments{
\item{myseed}{specifies seed for random assignment using set.seed().}

\item{N}{number of simulation runs.}

\item{smp.eff}{sample effect data, rows as probes, columns as samples.}

\item{norm.list}{a list of strings for normalization methods compared in the simulation study;
built-in normalization methods includes "NN", "QN", "MN", "VSN" for "No Normalization", "Quantile Normalization", "Median Normalization", "Variance Stablizing Normalization";
user can provide a list of normalization methods given the functions are supplied (also see norm.funcs).}

\item{class.list}{a list of strings for classification methods compared in the simulation study;
built-in classification methods are "PAM" and "LASSO" for "prediction analysis for microarrays" and "least absolute shrinkage and selection operator";
user can provide a list of classification methods given the correponding model-building and predicting functions are supplied (also see class.funcs and pred.funcs).}

\item{norm.funcs}{a list of strings for names of user-defined normalization method functions, in the order of norm.list excluding any built-in normalization methods.}

\item{class.funcs}{a list of strings for names of user-defined classification model-building functions, in the order of class.list excluding any built-in classification methods.}

\item{pred.funcs}{a list of strings for names of user-defined classification predicting functions, in the order of class.list excluding any built-in classification methods.}
}
\value{
benchmark analysis results with list of models built and internal and external misclassification error stored, also a list of assignment stored
}
\description{
Performs classification analysis on the uniformly-handled data by reassigning samples to training and test set in Qin et al. (see reference).
}
\examples{
\dontrun{
smp.eff <- estimate.smp.eff(r.data = r.data.pl)
ctrl.genes <- unique(rownames(r.data.pl))[grep("NC", unique(rownames(r.data.pl)))]
smp.eff.nc <- smp.eff[!rownames(smp.eff) \%in\% ctrl.genes, ]
uni.handled.results <- uni.handled.simulate(myseed = 1, N = 3,
                                            smp.eff = smp.eff.nc,
                                            norm.list = c("NN", "QN"),
                                            class.list = c("PAM", "LASSO"))
}
}
\references{
http://clincancerres.aacrjournals.org/content/20/13/3371.long
}
\keyword{simulation}

