% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs.norm.R
\name{vs.norm}
\alias{vs.norm}
\title{Variance stabalizing nomalization}
\usage{
vs.norm(train, test = NULL)
}
\arguments{
\item{train}{training data to be variance stabalizing normalized, rows as probes, columns as samples.}

\item{test}{test data to be frozen variance stabalizing normalized, rows as probes with equal number of rows as the training set, columns as samples.}
}
\value{
a list of two datasets, the normalized training set and the frozen normalied test set
}
\description{
Normalizes training dataset with vsn, stores the fitted vsn model from the training dataset as the reference to frozen variance stabalizng normalize test dataset.
}
\examples{
\dontrun{
set.seed(101)
group.id <- substr(colnames(non.r.data.pl), 7, 7)
train.ind <- colnames(non.r.data.pl)[c(sample(which(group.id == "E"), size = 64),
                               sample(which(group.id == "V"), size = 64))]
train.dat <- non.r.data.pl[, train.ind]
test.dat <- non.r.data.pl[, !colnames(non.r.data.pl) \%in\%train.ind]
data.vsn <- vs.norm(train = train.dat)
str(data.vsn)
data.vsn <- vs.norm(train = train.dat, test = test.dat)
str(data.vsn)
}
}
\references{
Wolfgang Huber, Anja von Heydebreck, Holger Sueltmann, Annemarie Poustka and Martin Vingron. Variance Stabilization Applied to Microarray Data Calibration and to the Quantification of Differential Expression. Bioinformatics 18, S96-S104 (2002).
}
\keyword{preprocess}

