% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.design.R
\name{blocking.design}
\alias{blocking.design}
\title{Blocking Design}
\usage{
blocking.design(seed, num.smp)
}
\arguments{
\item{seed}{an integer used to initialize a pseudorandom number generator.}

\item{num.smp}{number of arrays. It must be a multiple of 8.}
}
\value{
a vector of array IDs in the order of assigning to samples that are assumed to be sorted by sample group of interest
As a result, the first half of the array IDs are assigned to group 1 and the second half of the array IDs are assigned to group 2.
}
\description{
Assign arrays to samples with blocking by (8-plex Agilent) array slide.
}
\examples{
blocking.design(seed = 1, num.smp = 128)

}
\keyword{study.design}

