% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.design.R
\name{blocking.design}
\alias{blocking.design}
\title{Blocking Design}
\usage{
blocking.design(seed, num.smp)
}
\arguments{
\item{seed}{specifies seed for random assignment using set.seed().}

\item{num.smp}{number of samples.}
}
\value{
array-to-sample assignment, first half for group 1 (endometrial), second half for group 2 (ovarian)
}
\description{
Assigns arrays to samples with blocking design.
}
\examples{
blocking.design(seed = 1, num.smp = 128)

}
\keyword{study.design}

