% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.ary.eff.R
\name{estimate.ary.eff}
\alias{estimate.ary.eff}
\title{Estimated array effect}
\usage{
estimate.ary.eff(r.data, non.r.data)
}
\arguments{
\item{r.data}{randomized expression dataset, rows as probes, columns as samples.}

\item{non.r.data}{non-randomized expression dataset, rows as probes, columns as samples; must have same dimensions and same probe name as randomized data.}
}
\value{
an estimation of the array effect
}
\description{
Estimates array effect from taking the differences between the expressions of the non-randomized and the randomized data, matched by samples
}
\examples{
ary.eff <- estimate.ary.eff(r.data = r.data.pl, non.r.data = non.r.data.pl)
}
\keyword{data.setup}

