% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.smp.eff.R
\name{estimate.smp.eff}
\alias{estimate.smp.eff}
\title{Estimated Sample Effect}
\usage{
estimate.smp.eff(r.data)
}
\arguments{
\item{r.data}{randomized expression dataset, rows as probes, columns as samples.}
}
\value{
an estimation of the sample effect
}
\description{
Estimates sample effect from the expressions of the randomized data
}
\examples{
smp.eff <- estimate.smp.eff(r.data = r.data.pl)
}
\keyword{data.setup}

