% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.norm.R
\name{med.norm}
\alias{med.norm}
\title{Median normalization}
\usage{
med.norm(train, test = NULL)
}
\arguments{
\item{train}{training dataset to be median normalized.
The dataset must have rows as probes and columns as samples.}

\item{test}{test dataset to be frozen median normalized.
The dataset must have rows as probes and columns as samples.
The number of rows must equal to the number of rows in the training set.
By default, the test set is not specified (\code{test = NULL}) and
no frozen normalization will be performed.}
}
\value{
a list of two datasets:
\item{train.mn}{the normalized training set}
\item{test.fmn}{the frozen normalized test set, if test set is specified}
}
\description{
Normalize training dataset so that each array shares
a same median and store the median from the training dataset
as the reference to frozen median normalize test dataset.
}
\examples{
set.seed(101)
group.id <- substr(colnames(non.r.data.pl), 7, 7)
train.ind <- colnames(non.r.data.pl)[c(sample(which(group.id == "E"), size = 64),
                               sample(which(group.id == "V"), size = 64))]
train.dat <- non.r.data.pl[, train.ind]
test.dat <- non.r.data.pl[, !colnames(non.r.data.pl) \%in\%train.ind]
data.mn <- med.norm(train = train.dat)
str(data.mn)
data.mn <- med.norm(train = train.dat, test = test.dat)
str(data.mn)
}
\keyword{preprocess}

