% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.intcv.R
\name{pam.intcv}
\alias{pam.intcv}
\title{Nearest shrunken centroid through internal cross validation}
\usage{
pam.intcv(X, y, vt.k = NULL, n.k = 30, kfold = 5, folds = NULL, seed)
}
\arguments{
\item{X}{expression dataset to be trained, rows as probes, columns as samples.}

\item{y}{sample group corresponding to the data to be trained; must have the equal length as the number of samples as X.}

\item{vt.k}{custom-specified threshold list; default is NULL predetermined by the PAM package.}

\item{n.k}{number of threshold values desired; default is 30.}

\item{kfold}{number of folds for cross validation; default is 5}

\item{folds}{prespecifies samples to folds; default is NULL for no prespecification.}

\item{seed}{specifies seed for random assignment using set.seed().}
}
\value{
a PAM classifier
}
\description{
Builds a PAM classifier using internal cross validation, with 5-fold cross validation as the default.
}
\examples{
set.seed(101)
smp.eff <- estimate.smp.eff(r.data = r.data.pl)
ctrl.genes <- unique(rownames(r.data.pl))[grep("NC", unique(rownames(r.data.pl)))]
smp.eff.nc <- smp.eff[!rownames(smp.eff) \%in\% ctrl.genes, ]
group.id <- substr(colnames(smp.eff.nc), 7, 7)

smp.eff.train.ind <- colnames(smp.eff.nc)[c(sample(which(group.id == "E"), size = 64),
                                         sample(which(group.id == "V"), size = 64))]
smp.eff.nc.tr <- smp.eff.nc[, smp.eff.train.ind]

pam.int <- pam.intcv(X = smp.eff.nc.tr,
                     y = substr(colnames(smp.eff.nc.tr), 7, 7),
                     kfold = 5, seed = 1)
}
\keyword{classification}

