% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.simulate.R
\name{precision.simulate}
\alias{precision.simulate}
\title{Classification analysis of simulation study}
\usage{
precision.simulate(myseed, N, smp.eff.tr, smp.eff.te, ary.eff.tr, ary.eff.te,
  group.id.tr, group.id.te, design.list = c("CC+", "CC-", "PC+", "PC-"),
  norm.list = c("NN", "QN"), class.list = c("PAM", "LASSO"),
  batch.id = NULL, icombat = FALSE, isva = FALSE, iruv = FALSE,
  smp.eff.tr.ctrl = NULL, ary.eff.tr.ctrl = NULL, norm.funcs = NULL,
  class.funcs = NULL, pred.funcs = NULL)
}
\arguments{
\item{myseed}{specifies seed for random assignment using set.seed().}

\item{N}{number of simulation runs.}

\item{smp.eff.tr}{sample effect training data, rows as probes, columns as samples.}

\item{smp.eff.te}{sample effect test data, rows as probes, columns as samples; must have same number of probes and probe names as sample effect training data.}

\item{ary.eff.tr}{array effect training data, rows as probes, columns as samples; must have same dimensions and same probe name as sample effect training data.}

\item{ary.eff.te}{array effect test data, rows as probes, columns as samples; must have same dimensions and same probe name as array effect test data.}

\item{group.id.tr}{sample group ID of the sample effect training data; has to be "E" and "V".}

\item{group.id.te}{sample group ID of the sample effect test data; has to be "E" and "V".}

\item{design.list}{a list of strings for study designs compared in the simulation study;
built-in designs are "CC+", "CC-", "PC+", "PC-", "BLK", and "STR" for "Complete Confounding 1", "Complete Confounding 2", "Partial Confounding 1", "Partial Confounding 2", "Blocking", "Stratification" in Qin et al.}

\item{norm.list}{a list of strings for normalization methods compared in the simulation study;
build-in available normalization methods are "NN", "QN", "MN", "VSN" for "No Normalization", "Quantile Normalization", "Median Normalization", "Variance Stablizing Normalization";
user can provide a list of normalization methods given the functions are supplied (also see norm.funcs).}

\item{class.list}{a list of strings for classification methods compared in the simulation study;
built-in classification methods are "PAM" and "LASSO" for "prediction analysis for microarrays" and "least absolute shrinkage and selection operator";
user can provide a list of classification methods given the correponding model-building and predicting functions are supplied (also see class.funcs and pred.funcs).}

\item{batch.id}{a list of batch id by the number of batches when stratification study design is specified; default is NULL.}

\item{icombat}{indicator for combat adjustment; default is not to adjust (icombat = FALSE).}

\item{isva}{indicator for sva adjustment; default is not to adjust (isva = FALSE).}

\item{iruv}{indicator for RUV-4 adjustment; default is not to adjust (iruv = FALSE).}

\item{smp.eff.tr.ctrl}{negative-control gene sample effect data if iruv = TRUE, rows as probes, columns as samples;
must have same number of probes and probe names as non-control gene sample effect training data.}

\item{ary.eff.tr.ctrl}{negative-control gene array effect data if iruv = TRUE, rows as probes, columns as samples;
must have same number of probes and probe names as non-control gene array effect training data.}

\item{norm.funcs}{a list of strings for names of user-defined normalization method functions, in the order of norm.list excluding any built-in normalization methods.}

\item{class.funcs}{a list of strings for names of user-defined classification model-building functions, in the order of class.list excluding any built-in classification methods.}

\item{pred.funcs}{a list of strings for names of user-defined classification predicting functions, in the order of class.list excluding any built-in classification methods.}
}
\value{
simulated results with list of models built and internal and external misclassification error stored, also a list of assignment stored
}
\description{
Performs simulation study in Qin et al. (see reference).
}
\examples{
\dontrun{
set.seed(101)
smp.eff <- estimate.smp.eff(r.data = r.data.pl)
ary.eff <- estimate.ary.eff(r.data = r.data.pl,
                             non.r.data = non.r.data.pl)

ctrl.genes <- unique(rownames(r.data.pl))[grep("NC", unique(rownames(r.data.pl)))]

smp.eff.nc <- smp.eff[!rownames(smp.eff) \%in\% ctrl.genes, ]
ary.eff.nc <- ary.eff[!rownames(ary.eff) \%in\% ctrl.genes, ]

group.id <- substr(colnames(smp.eff.nc), 7, 7)

# randomly split sample effect data into training and test set with
# equal number of endometrial and ovarian samples
smp.eff.train.ind <- colnames(smp.eff.nc)[c(sample(which(group.id == "E"), size = 64),
                                          sample(which(group.id == "V"), size = 64))]
smp.eff.test.ind <- colnames(smp.eff.nc)[!colnames(smp.eff.nc) \%in\% smp.eff.train.ind]
smp.eff.train.test.split =
  list("tr" = smp.eff.train.ind,
       "te" = smp.eff.test.ind)

# non-randomly split array effect data into training and test set
ary.eff.train.test.split =
  list("tr" = c(1:64, 129:192),
       "te" = 65:128)

smp.eff.nc.tr <- smp.eff.nc[, smp.eff.train.ind]
smp.eff.nc.te <- smp.eff.nc[, smp.eff.test.ind]
ary.eff.nc.tr <- ary.eff.nc[, c(1:64, 129:192)]
ary.eff.nc.te <- ary.eff.nc[, 65:128]

# Simulation without batch adjustment
precision.results <- precision.simulate(myseed = 1, N = 3,
                                        smp.eff.tr = smp.eff.nc.tr,
                                        smp.eff.te = smp.eff.nc.te,
                                        ary.eff.tr = ary.eff.nc.tr,
                                        ary.eff.te = ary.eff.nc.te,
                                        group.id.tr = substr(colnames(smp.eff.nc.tr), 7, 7),
                                        group.id.te = substr(colnames(smp.eff.nc.te), 7, 7),
                                        design.list = c("PC-", "STR"),
                                        norm.list = c("NN", "QN"),
                                        class.list = c("PAM", "LASSO"),
                                        batch.id = list(1:40,
                                                        41:64,
                                                        (129:160) - 64,
                                                        (161:192) - 64))

# Simulation with RUV-4 batch adjustment
smp.eff.ctrl <- smp.eff[rownames(smp.eff) \%in\% ctrl.genes, ]
ary.eff.ctrl <- ary.eff[rownames(ary.eff) \%in\% ctrl.genes, ]

smp.eff.tr.ctrl <- smp.eff.ctrl[, smp.eff.train.test.split$tr]
ary.eff.tr.ctrl <- ary.eff.ctrl[, ary.eff.train.test.split$tr]

precision.ruv4.results <- precision.simulate(myseed = 1, N = 3,
                                             smp.eff.tr = smp.eff.nc.tr,
                                             smp.eff.te = smp.eff.nc.te,
                                             ary.eff.tr = ary.eff.nc.tr,
                                             ary.eff.te = ary.eff.nc.te,
                                             group.id.tr = substr(colnames(smp.eff.nc.tr), 7, 7),
                                             group.id.te = substr(colnames(smp.eff.nc.te), 7, 7),
                                             design.list = c("PC-", "STR"),
                                             norm.list = c("NN", "QN"),
                                             class.list = c("PAM", "LASSO"),
                                             batch.id = list(1:40,
                                                             41:64,
                                                             (129:160) - 64,
                                                             (161:192) - 64),
                                             iruv = TRUE,
                                             smp.eff.tr.ctrl = smp.eff.tr.ctrl,
                                             ary.eff.tr.ctrl = ary.eff.tr.ctrl)
}
}
\references{
http://clincancerres.aacrjournals.org/content/20/13/3371.long
}
\keyword{simulation}

