% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.signal.R
\name{reduce.signal}
\alias{reduce.signal}
\title{Biological signal reduction}
\usage{
reduce.signal(sample.effect, group.id, group.id.level = c("E", "V"),
  reduce.multiplier = 1/2, pbset.id = NULL)
}
\arguments{
\item{sample.effect}{the estimated sample effect dataset. The dataset must have rows as probes and columns as samples.
It can only take in probe-level dataset with a fixed number of probes per unique probe-set.}

\item{group.id}{a vector of sample-group labels for each sample of the estimated sample effect dataset.}

\item{group.id.level}{a vector of sample-group label level. It must have two and only two elements and the first element is the reference.
By default, \code{group.id.level = c("E", "V")}. That is in our study, we compare endometrial tumor samples to
ovarian tumor samples, with endometrial as our reference.}

\item{reduce.multiplier}{a multiplier specified to reduce between-sample-group signal by. By default, \code{reduce.multiplier = 1/2}.}

\item{pbset.id}{a vector of unique probe-set names. If it is not specified, it is the unique probe names of the dataset,
extracting from the row names.}
}
\value{
estimated sample effect data, with reduced biological signal
}
\description{
Reduce biological signal by decreasing the mean group difference between sample groups.
}
\examples{
sample.effect <- estimate.sample.effect(uhdata = uhdata.pl)
array.effect <- estimate.array.effect(uhdata = uhdata.pl,
                             nuhdata = nuhdata.pl)

ctrl.genes <- unique(rownames(uhdata.pl))[grep("NC", unique(rownames(uhdata.pl)))]

sample.effect.nc <- sample.effect[!rownames(sample.effect) \%in\% ctrl.genes, ]
array.effect.nc <- array.effect[!rownames(array.effect) \%in\% ctrl.genes, ]
group.id <- substr(colnames(sample.effect.nc), 7, 7)

redhalf.sample.effect.nc <- reduce.signal(sample.effect = sample.effect.nc,
                                    group.id = group.id,
                                    group.id.level = c("E", "V"),
                                    reduce.multiplier = 1/2)
}
\keyword{data.setup}

