% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs.norm.R
\name{vs.norm}
\alias{vs.norm}
\title{Variance stabilizing normalization}
\usage{
vs.norm(train, test = NULL)
}
\arguments{
\item{train}{training dataset to be variance stabilizing normalized. The dataset must have rows as probes and columns as samples.}

\item{test}{test dataset to be frozen variance stabilizing normalized. The dataset must have rows as probes and columns as samples.
The number of rows must equal to the number of rows in the training set.
By default, the test set is not specified (\code{test = NULL}) and no frozen normalization will be performed.}
}
\value{
a list of two datasets:
\item{train.mn}{the normalized training set}
\item{test.fmn}{the frozen normalized test set, if test set is specified}
}
\description{
Normalize training dataset with vsn and
store the fitted vsn model from the training dataset as the reference to frozen variance stabilizing normalize test dataset.
}
\examples{
\dontrun{
set.seed(101)
group.id <- substr(colnames(non.r.data.pl), 7, 7)
train.ind <- colnames(non.r.data.pl)[c(sample(which(group.id == "E"), size = 64),
                               sample(which(group.id == "V"), size = 64))]
train.dat <- non.r.data.pl[, train.ind]
test.dat <- non.r.data.pl[, !colnames(non.r.data.pl) \%in\%train.ind]
data.vsn <- vs.norm(train = train.dat)
str(data.vsn)
data.vsn <- vs.norm(train = train.dat, test = test.dat)
str(data.vsn)
}
}
\references{
Wolfgang Huber, Anja von Heydebreck, Holger Sueltmann, Annemarie Poustka and Martin Vingron.
Variance Stabilization Applied to Microarray Data Calibration and to the Quantification of Differential Expression.
Bioinformatics 18, S96-S104 (2002).
}
\keyword{preprocess}

